release v0.0.5
In addition to force-pushing a branch reference (deploy-staging, etc) also push tags so that there is a history of everything that is depl'd. By default it will let you use your `EDITOR` to set a deploy message.

Parameterize the remote, so that you can deploy to something other than `origin` if you'd like.